Refactored installApps configuration to update libertyConfigXml at create time
