+++ src/main/groovy/net/wasdev/wlp/gradle/plugins/tasks/InstallAppsTask.groovy
@@ -52,13 +52,9 @@ class InstallAppsTask extends AbstractServerTask {
 
     @TaskAction
     void installApps() {
-        if ((server.apps == null || server.apps.isEmpty()) && (server.dropins == null || server.dropins.isEmpty())) {
-            if (project.plugins.hasPlugin('war')) {
-                server.apps = [project.war]
-            } else if (project.plugins.hasPlugin('ear')) {
-                server.apps = [project.ear]
-            }
-        }
+
+        configureApps(project)
+
         if (server.apps != null && !server.apps.isEmpty()) {
             Tuple appsLists = splitAppList(server.apps)
             installMultipleApps(appsLists[0], 'apps')
@@ -75,13 +71,11 @@ class InstallAppsTask extends AbstractServerTask {
         if (applicationXml.hasChildElements()) {
             logger.warn("At least one application is not defined in the server configuration but the build file indicates it should be installed in the apps folder. Application configuration is being added to the target server configuration dropins folder by the plug-in.")
             applicationXml.writeApplicationXmlDocument(getServerDir(project))
-            writeServerPropertiesToXml(project)
         } else if (hasConfiguredApp(libertyConfigDropinsAppXml)) {
             logger.warn("At least one application is not defined in the server configuration but the build file indicates it should be installed in the apps folder. Liberty will use additional application configuration added to the the target server configuration dropins folder by the plug-in.")
         } else {
             if (libertyConfigDropinsAppXml.exists()){
                 libertyConfigDropinsAppXml.delete()
-                writeServerPropertiesToXml(project)
             }
         }
     }
@@ -95,41 +89,6 @@ class InstallAppsTask extends AbstractServerTask {
     private void installProjectArchive(Task task, String appsDir){
       Files.copy(task.archivePath.toPath(), new File(getServerDir(project), "/" + appsDir + "/" + getArchiveName(task)).toPath(), StandardCopyOption.REPLACE_EXISTING)
       validateAppConfig(getArchiveName(task), task.baseName, appsDir)
-    }
-
-    protected void validateAppConfig(String fileName, String artifactId, String dir) throws Exception {
-        String appsDir = dir
-        if (appsDir.equalsIgnoreCase('apps') && !isAppConfiguredInSourceServerXml(fileName)) {
-            applicationXml.createApplicationElement(fileName, artifactId)
-        }
-        else if (appsDir.equalsIgnoreCase('dropins') && isAppConfiguredInSourceServerXml(fileName)) {
-            throw new GradleException("The application, " + artifactId + ", is configured in the server.xml and the plug-in is configured to install the application in the dropins folder. A configured application must be installed to the apps folder.")
-        }
-    }
-
-    protected boolean isAppConfiguredInSourceServerXml(String fileName) {
-        boolean configured = false;
-        File serverConfigFile = new File(getServerDir(project), 'server.xml')
-        if (serverConfigFile != null && serverConfigFile.exists()) {
-            try {
-                ServerConfigDocument scd = new ServerConfigDocument(serverConfigFile, server.configDirectory, server.bootstrapPropertiesFile, server.bootstrapProperties, server.serverEnv)
-                if (scd != null && scd.getLocations().contains(fileName)) {
-                    logger.debug("Application configuration is found in server.xml : " + fileName)
-                    configured = true
-                }
-            }
-            catch (Exception e) {
-                logger.warn(e.getLocalizedMessage())
-            }
-        }
-        return configured
-    }
-
-    protected String getArchiveName(Task task){
-        if (server.stripVersion){
-            return task.baseName + "." + task.extension
-        }
-        return task.archiveName;
     }
 
     protected void installProject(Task task, String appsDir) throws Exception {
@@ -350,23 +309,6 @@ class InstallAppsTask extends AbstractServerTask {
         appFiles.each { File appFile ->
             installFromFile(appFile, appsDir)
         }
-    }
-
-    private Tuple splitAppList(List<Object> allApps) {
-        List<File> appFiles = new ArrayList<File>()
-        List<Task> appTasks = new ArrayList<Task>()
-
-        allApps.each { Object appObj ->
-            if (appObj instanceof Task) {
-                appTasks.add((Task)appObj)
-            } else if (appObj instanceof File) {
-                appFiles.add((File)appObj)
-            } else {
-                logger.warn('Application ' + appObj.getClass.name + ' is expressed as ' + appObj.toString() + ' which is not a supported input type. Define applications using Task or File objects.')
-            }
-        }
-
-        return new Tuple(appTasks, appFiles)
     }
 
     //Checks if there is an app configured in an existing configDropins application xml file
